+++ sdk/cosmos/azure-cosmos/src/samples/java/com/azure/cosmos/AsyncContainerCodeSnippets.java
@@ -4,10 +4,14 @@
 package com.azure.cosmos;
 
 import com.azure.cosmos.models.*;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import static com.azure.cosmos.ReadmeSamples.*;
 
 public class AsyncContainerCodeSnippets {
     private final CosmosAsyncClient cosmosAsyncClient = new CosmosClientBuilder()
@@ -147,4 +151,132 @@ public class AsyncContainerCodeSnippets {
         // END: com.azure.cosmos.CosmosAsyncContainer.readAllConflicts
     }
 
+    public void patchItemAsyncSample() {
+        Passenger passenger = new Passenger("carla.davis@outlook.com", "Carla Davis", "SEA", "IND");
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.patchItem
+        CosmosPatchOperations cosmosPatchOperations = CosmosPatchOperations.create();
+
+        cosmosPatchOperations
+            .add("/departure", "SEA")
+            .increment("/trips", 1);
+
+        cosmosAsyncContainer.patchItem(
+                passenger.getId(),
+                new PartitionKey(passenger.getId()),
+                cosmosPatchOperations,
+                Passenger.class)
+            .subscribe(response -> {
+                System.out.println(response);
+            }, throwable -> {
+                throwable.printStackTrace();
+            });
+        // END: com.azure.cosmos.CosmosAsyncContainer.patchItem
+    }
+
+    public void replaceItemAsyncSample() {
+        Passenger oldPassenger = new Passenger("carla.davis@outlook.com", "Carla Davis", "SEA", "IND");
+        Passenger newPassenger = new Passenger("carla.davis@outlook.com", "Carla Davis", "SEA", "IND");
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.replaceItem
+        cosmosAsyncContainer.replaceItem(
+                newPassenger,
+                oldPassenger.getId(),
+                new PartitionKey(oldPassenger.getId()),
+                new CosmosItemRequestOptions())
+            .subscribe(response -> {
+                System.out.println(response);
+            }, throwable -> {
+                throwable.printStackTrace();
+            });
+        // END: com.azure.cosmos.CosmosAsyncContainer.replaceItem
+    }
+
+    public void readAllItemsAsyncSample() {
+        String partitionKey = "partitionKey";
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.readAllItems
+        cosmosAsyncContainer
+            .readAllItems(new PartitionKey(partitionKey), Passenger.class)
+            .byPage(100)
+            .flatMap(passengerFeedResponse -> {
+                for (Passenger passenger : passengerFeedResponse.getResults()) {
+                    System.out.println(passenger);
+                }
+                return Flux.empty();
+            })
+            .subscribe();
+        // END: com.azure.cosmos.CosmosAsyncContainer.readAllItems
+    }
+
+    public void readManyAsyncSample() {
+        String passenger1Id = "item1";
+        String passenger2Id = "item1";
+
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.readMany
+        List<CosmosItemIdentity> itemIdentityList = new ArrayList<>();
+        itemIdentityList.add(new CosmosItemIdentity(new PartitionKey(passenger1Id), passenger1Id));
+        itemIdentityList.add(new CosmosItemIdentity(new PartitionKey(passenger2Id), passenger2Id));
+
+        cosmosAsyncContainer.readMany(itemIdentityList, Passenger.class)
+            .flatMap(passengerFeedResponse -> {
+                for (Passenger passenger : passengerFeedResponse.getResults()) {
+                    System.out.println(passenger);
+                }
+                return Mono.empty();
+            })
+            .subscribe();
+        // END: com.azure.cosmos.CosmosAsyncContainer.readMany
+    }
+
+    public void queryChangeFeedSample() {
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.queryChangeFeed
+        CosmosChangeFeedRequestOptions options = CosmosChangeFeedRequestOptions
+            .createForProcessingFromNow(FeedRange.forFullRange())
+            .allVersionsAndDeletes();
+
+        cosmosAsyncContainer.queryChangeFeed(options, Passenger.class)
+            .byPage()
+            .flatMap(passengerFeedResponse -> {
+                for (Passenger passenger : passengerFeedResponse.getResults()) {
+                    System.out.println(passenger);
+                }
+                return Flux.empty();
+            })
+            .subscribe();
+        // END: com.azure.cosmos.CosmosAsyncContainer.queryChangeFeed
+    }
+
+    public void queryItemsAsyncSample() {
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.queryItems
+        CosmosQueryRequestOptions options = new CosmosQueryRequestOptions();
+        String query = "SELECT * FROM Passenger WHERE Passenger.departure IN ('SEA', 'IND')";
+        cosmosAsyncContainer.queryItems(query, options, Passenger.class)
+            .byPage()
+            .flatMap(passengerFeedResponse -> {
+                for (Passenger passenger : passengerFeedResponse.getResults()) {
+                    System.out.println(passenger);
+                }
+                return Flux.empty();
+            })
+            .subscribe();
+        // END: com.azure.cosmos.CosmosAsyncContainer.queryItems
+    }
+
+    public void queryItemsAsyncSqlQuerySpecSample() {
+        // BEGIN: com.azure.cosmos.CosmosAsyncContainer.SqlQuerySpec.queryItems
+        CosmosQueryRequestOptions options = new CosmosQueryRequestOptions();
+
+        String query = "SELECT * FROM Passenger p WHERE (p.departure = @departure)";
+        List<SqlParameter> parameters = Collections.singletonList(new SqlParameter("@departure", "SEA"));
+        SqlQuerySpec sqlQuerySpec = new SqlQuerySpec(query, parameters);
+
+        cosmosAsyncContainer.queryItems(sqlQuerySpec, options, Passenger.class)
+            .byPage()
+            .flatMap(passengerFeedResponse -> {
+                for (Passenger passenger : passengerFeedResponse.getResults()) {
+                    System.out.println(passenger);
+                }
+                return Flux.empty();
+            })
+            .subscribe();
+        // END: com.azure.cosmos.CosmosAsyncContainer.SqlQuerySpec.queryItems
+    }
 }